refactor

+++ src/features/users/users-router.ts
@@ -7,6 +7,7 @@ import {QueryUserModel} from "../../models/usersModels";
 import {usersQueryRepository} from "./users-query-repository";
 import {userValidator} from "./middlewaresUsers";
 import {usersService} from "./users-service";
+import {isValidIdFromObjectId} from "../../helpers/checkValidIdfromObjectId";
 
 export const usersRouter = express.Router();
 
@@ -34,7 +35,13 @@ usersRouter.post("/", authMiddleware, ...userValidator, sendAccumulatedErrorsMid
     })
 
 //delete user by id
-usersRouter.delete("/:id", authMiddleware, async (req: Request, res: Response) => {
+usersRouter.delete("/:id", authMiddleware, sendAccumulatedErrorsMiddleware, async (req: Request, res: Response) => {
+    const id = req.params.id
+    const isValidId = isValidIdFromObjectId(id)
+    if (!isValidId) {
+        res.sendStatus(HTTP_STATUS.NOT_FOUND)
+        return
+    }
     const isDeleted = await usersService.deleteUserById(req.params.id)
     if (isDeleted) {
         res.sendStatus(HTTP_STATUS.NO_CONTENT)
